backend/requirements.txt
django>=4.2
djangorestframework>=3.14
supabase>=2.0,<2.10
google-generativeai>=0.3
azure-cognitiveservices-vision-computervision>=0.9
requests>=2.31
gunicorn>=21.2
Pillow>=10.0
django-cors-headers>=4.3
python-dotenv>=1.0
hypothesis>=6.0
PyJWT>=2.8
